fix: remove bug when loading styles
