--- src/app/home/home.module.css
@@ -1,5 +1,0 @@
-.layout {
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
-}
