--- src/app/home/page.tsx
@@ -1,16 +1,0 @@
-import { FooterHome } from "~/components/footer-home/footer-home";
-import { HomeMain } from "~/components/home-main/home-main";
-import { HomeHeader } from "~/components/home-header/home-header";
-import styles from "./home.module.css";
-
-export default function Home() {
-  return (
-    <div className={styles.layout}>
-      <HomeHeader />
-
-      <HomeMain />
-
-      <FooterHome />
-    </div>
-  );
-}
